+++ src/api/user/user.module.ts
@@ -3,9 +3,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { User } from '@src/entity/user.entity';
 
-import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { UserLib } from './service/user.lib';
+import { UserService } from './service/user.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
